+++ src/main/java/se/bjurr/prnfs/settings/PrnfsNotificationBuilder.java
@@ -8,6 +8,9 @@ import java.util.List;
 import se.bjurr.prnfs.listener.PrnfsPullRequestAction;
 
 public class PrnfsNotificationBuilder {
+ public static final String YES = "YES";
+ public static final String NO = "NO";
+
  public static PrnfsNotificationBuilder prnfsNotificationBuilder() {
   return new PrnfsNotificationBuilder();
  }
@@ -20,12 +23,18 @@ public class PrnfsNotificationBuilder {
  private String filterString;
  private String method;
  private String postContent;
+ private final List<Header> headers = newArrayList();
+ private String proxyUser;
+ private String proxyPassword;
+ private String proxyServer;
+ private String proxyPort;
 
  private PrnfsNotificationBuilder() {
  }
 
  public PrnfsNotification build() throws ValidationException {
-  return new PrnfsNotification(triggers, url, user, password, filterString, filterRegexp, method, postContent);
+  return new PrnfsNotification(triggers, url, user, password, filterString, filterRegexp, method, postContent, headers,
+    proxyUser, proxyPassword, proxyServer, proxyPort);
  }
 
  public PrnfsNotificationBuilder withPassword(String password) {
@@ -65,6 +74,31 @@ public class PrnfsNotificationBuilder {
 
  public PrnfsNotificationBuilder withPostContent(String postContent) {
   this.postContent = checkNotNull(postContent);
+  return this;
+ }
+
+ public PrnfsNotificationBuilder withHeader(String name, String value) {
+  headers.add(new Header(checkNotNull(name), checkNotNull(value)));
+  return this;
+ }
+
+ public PrnfsNotificationBuilder withProxyServer(String s) {
+  this.proxyServer = checkNotNull(s);
+  return this;
+ }
+
+ public PrnfsNotificationBuilder withProxyPort(String s) {
+  this.proxyPort = checkNotNull(s);
+  return this;
+ }
+
+ public PrnfsNotificationBuilder withProxyUser(String s) {
+  this.proxyUser = checkNotNull(s);
+  return this;
+ }
+
+ public PrnfsNotificationBuilder withProxyPassword(String s) {
+  this.proxyPassword = checkNotNull(s);
   return this;
  }
 }